Fix a bug in plugin reload + Add memory profiler
With the plugin structure and implementation of `unload_extensions`, we
need to manually reload our plugins. This could lead to memory leak
because we're not cleanly removing our plugin (or at least that's what
I think). I tested with `mprof` for 100000 reloads operation with the
About plugin, and while there is a memory increase, it's only 7MB, so
it shouldn't be a big concern.

**Shouldn't**, because I'm not knowledgeable about Python's import
system, and this is a very dodgy fix. If you're reading this after
`git bisect` and/or `git blame` due to memory leak, the fix implemented
here could be the culprit.